kernel-fsl-sdk-v1.7: patches to work with powerpc64 binutils 2.28
Compiling the Linux kernel for powerpc64 from fsl-sdk-v1.7 with
gcc-4.9.4 and binutils 2.28 requires back porting a few patches.

This version of Linux is used by some powerpc64 targets.

One patch in this set comes directly from the upstream Linux kernel.

The second patch solves the same type of problem as the first patch,
but the structure of the surrounding code is quite different
upstream.  Thus this patch is custom and hand written.

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>