i2ctools: set I2CTOOLS_SYSEEPROM=no as the default
Using the system EEPROM support in i2ctools has been deprecated since
2014.  The accepted method is to use the support added to busybox with
this commit:

  commit fc76004dca594a5f4f12a7905474f77c67e5effb
  Author: david_yang <david_yang@accton.com>
  Date:   Thu Sep 18 09:33:01 2014 +0800
   
      Add to support onie-syseeprom command in busybox

The legacy default for x86_64 machines is I2CTOOLS_SYSEEPROM=yes,
which is painful as all new machines have to explicitly set it to "no"
as they are using the busybox implementation.

This is dumb. The time come to change the default of this option to
no.

This patch also goes through all the legacy machines using the
i2ctools syseeprom implementation, which were relying on the old
default value of "yes", and explicitly sets the option to "yes" in the
corresponding machine.make file.

Closes: #565
Signed-off-by: Curt Brune <curt@cumulusnetworks.com>