Add ".bin" extension to default waterfall file names
In the field a number of user errors have been reported where the
setting of the default ONIE waterfall image is incorrect.  In most
cases the user has mistakenly appended ".bin" to the simple base name
"onie-installer".

The error is so trivially simple that most will miss it, which
prevents ONIE from finding installers.  This is especially common when
performing USB-based installations.

When the file extension is not detected, the USB stick then has to be
mounted manually and manually installed. This process can make Open
Networking look harder than it is.

This patch adds infrastructure for appending any number of suffixes to
the waterfall image names.  This infrastructure is then used to append
".bin" to the default names.

Closes: #567
Signed-off-by: Curt Brune <curt@cumulusnetworks.com>