During ONIE update, do not modify EEPROM TLV 0x28
A historical artifact in the ONIE installer is to update the platform
name (TLV 0x28) in the system EEPROM as part of the upgrade.

That operation is not needed as the platform name should be static
across upgrades.

This patch removes the code from the ONIE installer that was updating
EEPROM TLV 0x28 during ONIE upgrades.

Closes: #572
Signed-off-by: Curt Brune <curt@cumulusnetworks.com>