Identifying machines with common CPU modules
This patch adds support for machines that use a common CPU module.
These modules provide a common CPU, memory and storage system for
switching platforms. The CPU module can be plugged into a variety of
base boards where the switching elements reside.

The idea is to have a single ONIE "machine" for the CPU module and
detect at run time what the baseboard is.

The platform identification originates from /etc/machine.conf, a
portion of which is now derived at run time.  Along with this a new
configuration variable, "onie_build_machine" is introduced.

/etc/machine.conf
=================

Previously this file was completely built at compile time.  This file
is now a small wrapper script that sources two new files,
/etc/machine-build.conf and /etc/machine-live.conf.

/etc/machine-build.conf -- this file is completely built at compile
time and only contains information known at build time.  This file
introduces a new configuration variable, "onie_build_machine", which
is the ONIE machine specified at compile time.  For example this would
be the ONIE machine name of the common CPU module.

/etc/machine-live.conf -- this file is built at run time.  A boot time
init script sources a platform specific file if it exists, executes
the gen_live_config() function and stores the output in
/etc/machine-live.conf.  By default this function outputs nothing, but
a platform can override it produce a runtime ONIE machine name.

Using this mechanism, a machine can redefine 'onie_machine' and
'onie_switch_asic' at run time.  To do this, a machine defines a small
script in the source tree at
machine/<vendor>/rootconf/sysroot-lib-onie/gen-config-platform, which
include a definition of the gen_live_config() function.

If a machine does not define gen-config-platform,
i.e. /etc/machine-live.conf is empty, then the contents of
'onie_build_machine' is used to set 'onie_machine'.  This is
backwardly compatible with how ONIE worked previously.

Updating ONIE
=============

When ONIE updates itself, it will now check that the running
'onie_build_machine' matches the 'onie_build_machine' of the proposed
update image.  Previously it checked that 'onie_machine' matched
between runtime and the image.

Closes: #570
Signed-off-by: Curt Brune <curt@cumulusnetworks.com>