coreboot: build target grub-i386-pc for .ISO installer
In order to build an .ISO ONIE installer image we need the
grub-i386-pc components.

The ONIE build system builds the GRUB components twice.  Once using
the cross-compiler for the target architecture for use at runtime on
the target system.  And again using the host compiler for use by the
build system at build time.

On top of that, the cross-compiled components are built for a specific
firmware type: x86_64-efi, i386-pc or coreboot.

In ONIE versions 2016.11br and prior, the building of the .ISO ONIE
installer used the "host" version of the GRUB components, which always
included the i386-pc.

Commit de342ca1f8e5966 changed the .ISO build from using the 'host'
components to using the 'target' components:

  commit de342ca1f8e5966d0768d2319c94dd34f33eb98a
  Author: Curt Brune <curt@cumulusnetworks.com>
  Date:   Mon Nov 28 10:13:14 2016 -0800

      grub-x86_64-efi: fix for PXE_EFI64_ENABLE configurations

Unfortunately this broke the coreboot firmware type, since the target
version of grub-i386-pc is not built in this case.

The fix here is to also build grub-i386-pc for the target when the
firmware type is coreboot.

Closes: #487
Signed-off-by: Curt Brune <curt@cumulusnetworks.com>