kvm_x86_64: add demonstration secure boot keys and certificates
This patch enables the secure boot build for this platform.

Also, for demonstration purposes only, this patch adds a number of RSA
private keys and x509 certificates for testing.

For a hypothetical hardware vendor create private keys and
certificates for:

  - UEFI PK  -- platform owner key
  - UEFI KEK -- key exchange key
  - UEFI db  -- signature database

For a hypothetical software vendor create private keys and
certificates for:

  - UEFI KEK -- key exchange key
  - UEFI db  -- signature database

For a hypothetical ONIE vendor create private keys and certificates
for the embedded SHIM key pair.

For initializing the KEK on a blank EFI system, create a KEK EFI
signature list (ESL), containing hw-vendor and sw-vendor KEKs, signed
by the hw-vendor PK.

For initializing the db on a blank EFI system, create a DB EFI
signature list (ESL), containing hw-vendor and sw-vendor DBs, signed
by PK.

For production machines, the private keys would never be checked into
the ONIE code repository.

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>