pkg-config: prepend $(DEV_SYSROOT) to responses
Update the $(ONIE_PKG_CONFIG) environment variables to also set
PKG_CONFIG_SYSROOT_DIR.  This value is prepended to the results of -I
and -L.  See man pkg-config(1) for details.

The ONIE build system has been a little inconsistent in the past,
sometimes add $(DEV_SYSROOT) to configure's --prefix argument,
sometimes not.  The correct thing to do is:

1. add $(DEV_SYSROOT) to PKG_CONFIG_SYSROOT_DIR
2. set --prefix=/usr when running ./configure for a package
3. add DESTDIR=$(DEV_SYSROOT) to 'make install' a package

This patch is taking care of item 1.

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>