pkg-config: configure packages' --prefix properly
The ONIE build system has been a little inconsistent in the past,
sometimes adding $(DEV_SYSROOT) to configure's --prefix argument,
sometimes not.  The correct thing to do is:

1. set --prefix=/usr when running ./configure for a package
2. add DESTDIR=$(DEV_SYSROOT) to 'make install' a package

This patch updates a number a packages to use configure correctly.

With this change the LVM package no longer requires patching, as the
patching was working around the fact we were using pkg-config
incorrectly.

When configured correctly, the util-linux package puts libraries in
slightly different places than before.  This patch also accommodates
these new locations.

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>