Add build time platform information to GRUB messages
Commit 3cbcb9ef32c changed the way some of the $onie_XXX variables are
created and used, in particular $onie_platform is now computed are
runtime:

commit 3cbcb9ef32cdeffb2e036565271cbd010cf1f005
Author: Curt Brune <curt@cumulusnetworks.com>
Date:   Wed Aug 16 16:02:29 2017 -0700

    Identifying machines with common CPU modules

At runtime we now synthesize $onie_platform from runtime information
if available, otherwise we use the build time information.

For GRUB, this had the effect of making $onie_platform blank and GRUB
would print a blank platform string.

For GRUB we have no runtime info, so the best we can do is use the
build time info.

This patch echos the "build time" platform information when GRUB loads
and boots a kernel.

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>