kernel: updates for secure boot
When targeting secure boot, update the build system to sign the Linux
kernel with the ONIE_VENDOR_SECRET_KEY.

Also add MODULE_SIG_KEY_SRCPREFIX to the kernel build command, in
order to include the kernel owner's signing keys into the kernel.
This can be used by the kernel at runtime to verify signatures on
kernel modules and other files.

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>