onie-fwpkg: correct preservation of update records
The onie-fwpkg command is supposed to display the ONIE update (and
firmware update) history on the box.

During ONIE updates the entire ONIE data directory, which includes the
location of the ONIE update records, is reformatted.  During this
process the update records are first preserved and then, after the
reformatting, restored to the previous location.

However, the restore logic has a bug.

This patch fixes the bug, so that the restoration of update records
works as intended.

Closes: #679
Signed-off-by: Curt Brune <curt@cumulusnetworks.com>