gcc-4.9.2: enable binutils plugin support
The newly updated efivar-30 component relies on LTO (link time
optimizer) support.  LTO support for the gcc-4.9.2 toolchain requires
that the underlying binutils (ar) be build with PLUGIN support.

This patch enables binutil plugin support for the gcc-4.9.2 x86_64
architecture only.  The x86_64 architecture is the only architecture
we have that uses both gcc-4.9.2 and efivar.

Closes: #695
Signed-off-by: Curt Brune <curt@cumulusnetworks.com>