grub: make setting debug kernel arguments easier
Since we call a GRUB shell function to generate the grub menu entries,
it is difficult to use the GRUB menu "edit" feature to append extra
args to the kernel.

For example, before this change, when editting the "Rescue" entry you
would see this:

            GNU GRUB  version 2.02

 +----------------------------------------------+
 |setparams 'ONIE: Rescue'                      |
 |                                              |
 |  onie_rescue                                 |
 +----------------------------------------------+

No clue how to add extra args with that definition.

This patch adds an empty "onie_debugargs=" setting to every GRUB menu
item.  To temporarily add additional kernel args, add values to
onie_debugargs.

After this change, the editting screen looks like

            GNU GRUB  version 2.02

 +----------------------------------------------+
 |setparams 'ONIE: Rescue'                      |
 |                                              |
 |  onie_debugargs=""                           |
 |  onie_rescue                                 |

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>