onie-boot-mode: set a persistent fallback boot option
This patch adds an "-f" option to the "onie-boot-mode" command to set
a persistent fallback ONIE boot mode.  The fallback option is used
when no default boot mode is set.

The default GRUB boot entry is controlled by two persistent grub
environment variables.	If neither variable is set, then the first
boot entry menu is used, which is "install" mode.

In priority order:

- onie_mode	-- highest priority
- onie_fallback -- only used if onie_mode is not set

The "onie_mode" variable is what onie-boot-mode has historically set
by default.  This is used by a NOS to command the boot mode into
"rescue" mode for example.  The old behavior here is not changing.

The default mode specified by "onie_mode", however, is not necessarily
permanent.  For example "rescue" is a one-shot, good for one boot
only.

The idea here is to add a new variable, called "onie_fallback", that
is persistent.  Nothing clears this variable automatically, except an
explicit call to onie-boot-mode.

The intended use case is a NOS installer would set the fallback to
"rescue" at the end of install a NOS.  That way if a user stumbles
into the ONIE GRUB menu by accident the default entry will be
"rescue", instead of the more damaging "install".

Closes: #690
Signed-off-by: Curt Brune <curt@cumulusnetworks.com>