onie-nos-mode: make callable from onie tools directory
The onie-nos-mode utility is included in the persistent ONIE tools
directory, /mnt/onie-boot/onie/tools/bin, so that a NOS can directly
access it.

The initial implementation did not account for the fact that
onie-nos-mode also calls out to the 'onie-boot-mode' command.  When
called from the NOS, onie-boot-mode is not in the $PATH variable.

This patch uses the assumption that onie-boot-mode and onie-nos-mode
both resided in the same directory to call onie-boot-mode directly.

Fixes: cf6e91cda5c5 ("NOS mode: a persistent NOS mode boot option")
Signed-off-by: Curt Brune <curt@cumulusnetworks.com>