build: add helper script to update machine kernel versions
This patch adds a script that helps ONIE maintainers move the default
kernel version forward for all the machines.  The script updates the
machine.make for the relevant machines and creates a git commit.

Step 1 - configure the script for the old kernel default version and
the new kernel default version.  Edit these lines:

  CURRENT_MAJOR=4.9
  CURRENT_MINOR=95
  NEW_MINOR=113

In this example we are moving from 4.9.95 to 4.9.113.

Step 2 - run the script from the top-level directory, like this:

  build-08:~/onie/$ ./build-config/scripts/onie-src-update-kernel-version

Step 3 - inspect the output, look at 'git status'.

Step 4 - push the result to upstream

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>