onie-syseeprom: Allow multiple vendor extension TLVs
This patch modifies the behavior of the onie-syseeprom -s (--set) command. If a
TLV type code is specified which is allowed to appear multiple times in the
EEPROM, then the TLV, if present, is not removed from the EEPROM before the new
TLV value is added to the EEPROM. Currently, the only TLV type code which can
appear multiple times within the EEPROM is the Vendor Extension (0xFD). Like all
other TLVs added with the -s option, vendor extension TLVs are added at the end
of the EEPROM, just before the CRC TLV. Deleting a vendor extension TLV with the
"onie-syseeprom -s 0xfd" (no value) command will remove the first vendor
extension TLV in the EEPROM. This command can be used multiple times to remove
all vendor extension TLVs.

Testing has been done on an Accton 5712 platform by adding/deleting/replacing
both vendor extension and MAC address TLVs.

Signed-off-by: Scott Emery <scotte@cumulusnetworks.com>
Closes: #767