git-stats: scripts to generate ONIE release statistics
As part of making an ONIE release, I generate git repo statistics.
Finally moving these scripts and config files into the public repo for
all to use.

The bulk of the work is done by the gitdm script, which is courtesy
Jonathan Corbet <corbet@lwn.net>.  The ONIE specific part is in the
gitdm-config directory.

Signed-off-by: Curt Brune <curt@cumulusnetworks.com>