Alphanetworks STX-60D0-062F: Use acpi --no-ebda in ISO recovery menu
In the STX-60D0-062F, grub's "acpi" command cannot copy intact EBDA.
(Extended BIOS Data Area) if the EBDA size of BIOS is over 1K.
This command causes system hang.
Therefore the "acpi" command can only be executed with the
"--no-ebda" parameter.

This patch modifies the grub configuration of the recovery ISO prior to the
ISO file's creation, and the original grub-iso.cfg will be restored
after the ISO file is created.

Signed-off-by: TalisHsieh <talisych@gmail.com>
Signed-off-by: Alex Doyle <adoyle@cumulusnetworks.com>