Support system wide download cache at /var/cache/onie/download
The problem:

ONIE downloads specific versions of packages (most as compressed
tarfiles ) that get used in build. These packages are checksummed
against values stored in git, and are stored in onie/build/download.
These files are not removed on clean operations, although the compiled
versions of them are, so once created, onie/build/downloads is
fairly static, unless another ONIE build target is built, in which case
it may have to download additional packages, or different versions
of ones that it already has.

On a first-time checkout of the ONIE source code, this directory
gets populated by makefiles using ./scripts/fetch-package.

This script will iterate through a list of URLs to download the
source code. It starts with mirror.opencompute.org/onie, and
will then try other URLS, if any.

As some of the alternate URLs are old, they are no longer valid,
and mirror.opencompute.org/onie becomes the only reliable way to get them.

If something happens to this website, or the user has limited
Internet access, building ONIE becomes pretty much impossible.

...and it was the website going offline for a few days prompted this fix.

The solution:

The scripts/fetch-package script, as a last resort, will now
look for a system wide cache at /var/cache/onie/download.

Setting up this cache is entirely the responsibility of the
user, as it is not needed under normal circumstances. However if:

1 - Internet connectivity is, slow, intermittent, or non existent.
2 - or something happens to the mirror server.

...it might be nice to have a local copy of these files that
never change.

As of 10/31/19, the entire mirror takes up about 2GB of space, and
most users will only need a subset of these packages, so it
may be trimmed for individual use.

To set up the local copy:

One can either build all ONIE targets that one is interested in,
and copy the resulting onie/build/download folder to
/var/cache/onie/download

Or:

mkdir /var/cache/onie/download

...and poplulate it with the entire mirror using:

wget --recursive --cut-dirs=2 --no-host-directories \
 --no-parent --reject "index.html" http://mirror.opencompute.org/onie/

This was originally brought up in:
Bug URL: https://github.com/opencomputeproject/onie/issues/870

Fixes:  #870   ONIE build process missing packages
Signed-off-by: Alex Doyle <adoyle@cumulusnetworks.com>
Signed-off-by: adoyle <adoyle@cumulusnetworks.com>