Update of gh-pages documentation.
This commit has a number of changes as it was built
with a newer version of Sphinx - 1.7.9-1~bpo9+1, vs
the 1.6.7 version from Ubuntu 18.04.

The build change was introduced by building the entire
release inside a Debian 9 based Docker container, which
was done to maintain a consistent build/publication
environment for ONIE

Spot checks of the resulting documentation seem OK,
and copying the whole directory to a web server for
testing of the search functions and browsing showed that
everything seemed to be functioning correctly.

Signed-off-by: Alex Doyle adoyle@cumulusnetworks.com