Use local download cache with ONIE_USE_SYSTEM_DOWNLOAD_CACHE=TRUE
Setting ONIE_USE_SYSTEM_DOWNLOAD_CACHE=TRUE when invokging a build
will export that variable to be interpreted by scripts/fetch-package .

If set = TRUE, the fetch-package script will look under
/var/cache/onie/download for packages it would otherwise pull from
http://mirror.opencompute.org/onie.

The presumption is that the user would have set up this directory
ahead of time, either by copying files downloaded for the targets
they are building, or, in the case of a maintainer who has to
rebuild everything with some frequency, just downloading everything
from the OpenCompute mirror with:

 wget --recursive --cut-dirs=2 --no-host-directories \
  --no-parent --reject "index.html" http://mirror.opencompute.org/onie/

Note that the script will fail if a package is not present. This
is a 'feature' that will let the user know their cache is incomplete.

Signed-off-by: Alex Doyle <adoyle@cumulusnetworks.com>