Grub 2.04 - upgrade with Debian patch set
This upgrades GRUB to version 2.04, with a patch set from Debian 10.
Some patches may not be necessary, but the existing configuration
builds and runs with legacy and UEFI BIOS, and was tested with
the current KVM machine build for x86.

Other architectures have not been tested,but this is a starting
point for that.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>
Signed-off-by: Alex Doyle <alexddoyle@gmail.com>