Set crosstool<arch>.config CT_LINUX_VERSION from makefile LINUX_VERSION
The new crosstool config for uclbc requires that the version of the
Linux kernel being built be set in a few places. This is a change
from how things used to build.

As a build target's machine.make file can set this value with the
variable LINUX_VERSION, the value cannot be hardcoded in to the
crosstool<arch>.config, but needs to be set at build time.

This change leaves the config file's:
CT_LINUX_VERSION
...values unconfigured until the xtools.make makefile fragment can
apply the kernel values to the crosstool<arch>.config file that has
been copied to the build area.

Tested with successful builds of the Mellanox ONIE (4.9.95 kernel)
and the KVM (4.19.143 kernel)

Signed-off-by: Alex Doyle <adoyle@nvidia.com>
Signed-off-by: Alex Doyle <alexddoyle@gmail.com>