Upgrade crosstool-ng components and x86 config file
This updates the crosstool.x86_64.config file to a cleaner
configuration made with the crosstool-ng config upgrade
scripts, as well as upgrades the crosstool-ng components:

 isl  to 0.20
 mpc  to 1.1.0
 mpfr to 4.1.0

While these component versions are newer, they are not the
most recent as of this writing. However these versions were
explicitly referenced in the crosstool-ng configuration, and
were chosen for stability.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>
Signed-off-by: Alex Doyle <alexddoyle@gmail.com>