onie-build-targets.json - update armv8a platforms
This updates the onie-build-targets.json file to properly
identify a few armv8a platforms that had previously been
mis-classified as amd64.

armv8a is being used for the architecture type, as it
reflects the value ONIE_ARCH is set to in the machine.make
makefile fragment.

Thanks to Michael Shych at Nvidia for pointing this out.

Testing:
--------
 jq '.' build-config/scripts/onie-build-targets.json
 ...parsed the file with no errors.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>