Emulation support script for KVM
This adds onie-vm.sh, its function library onie-vm.lib,
 and README-onie-vm.md in a new emulation directory.

The onie-vm.sh script makes it easier to run ONIE's emulation
machines by handling the mechanics of setting up the virtual
hard drive, selecting BIOS type, defining ports to use on
the system, etc.

It is also designed to function independently of the ONIE build
area, so while it defaults to looking for a ../build directory,
paths can be supplied to the components being run.

This allows it to be run by users without having a full build
environment, as would be the case where QA engineers might want
a virtual system to write tests against while they wait for access
to new hardware.

Currently this works by default with the kvm-x86-64 system, but
can be expanded out to other architectures and machines, provided
the kernels of those machines are configured to support virtual
drives (/dev/vda, for example.)

Signed-off-by: Alex Doyle <adoyle@nvidia.com>