Add infrastructure for managing keys and signing.
This introduces makefile changes to clarify signing key uses.

As key management will be implemented on a per-vendor basis,
an ONIE build required some changes to cleanly support
unique key locations, and to provide a working example of this
using the kvm_x86_64 virtual machine.

The ./emulation directory provides a way to run a preconfigured
 kvm_x86_64 instance with a virtual USB drive to store files.

The ./encryption directory provides scripts to generate signing
 keys to demonstrate how a the keys from  hardware, software,
 and ONIE vendors can interact to create a valid Secure Boot
 configuration. It is integrated with the ./emulation directory
 structure so that it can replicate the experience of copying
 keys to be installed on a target's UEFI BIOS (via the virtual
 USB drive)

These changes integrate the aforementioned key management
 with the make system to:
 - Generate example keys.
 - Generate the self signed shim (if needed).
 - Clearly define which keys are getting used where.
 - Provide those keys to the emulated run time environment.

With that in mind, this introduces:

machine/kvm_x86_64/machine-security.make
 ...which is an example of how a machine specific makefile
 will contain all of the key paths and security settings
 for a build.
 It acts in a similar capacity to the machine.make file.
 NOTE: that the machine.make file defines:
  MACHINE_SECURITY_MAKEFILE
  ...which is the path to the security file to use.
  This can be overridden to distinguish between different
  build contexts, such as developer builds, and official
  releases.

build-config/make/signing-keys.make
 - Which will generate keys, and display the values of
   all Makefile variables used from the build target's
   machine-security.make file

The commit also updates Makefile, kernel.make, and shim.make
 to be aware of the variables set by the machine-security.make.

For example,

The signed shim is typically not part of the build/sign process, and
only comes in at image assembly time, as, in a real-world deployment
it would have been built and sent off to a third party to be signed
(Microsoft) or, in the case of a hardware manufacturer, signed by their
release key.

Point being that the ONIE build should pull the signed shim from a
'safe place' to use in the image.
For demonstration purposes, if ONIE signs the SHIM, the 'safe place'
defaults to being under:
  onie/encryption/machines/<build target>/safe-place,
 and these makefile changes will put it there.

Build Test:
-----------
 1 - make MACHINE=kvm_x86-64 signing-keys-generate
 2 - make MACHINE=kvm_x86-64 signing-keys-install
 3 - make MACHINE=kvm_x86-64 -j4 shim-self-sign
 4 - make MACHINE=kvm_x86-64 -j4 all

Where:
 1 - Create all possible demonstration keys under
     ./encryption/machines/<build target>/keys
 2 - Copy those keys to the ./emulation directory's
      virtual USB drive so they can be added to the
	  UEFI BIOS at run time
 3 - Create a self signed shim. Store it in
      SAFE_PLACE_DIRECTORY, as defined in machine-security.make
 4 - Build the rest of ONIE

Emulation test:
---------------
1 - cd onie/emulation
2 - ./onie-vm.sh --m-bios-uefi --m-usb-drive  --m-embed-onie \
  --m-onie-iso ../build/images/onie-recovery-x86_64-kvm_x86_64-r0.iso \
  --m-hd-clean.sh --m-bios-uefi --m-usb-drive
3 - (in separate window)  telnet localhost 9300
 ...and follow steps for embedding ONIE on qcow2 virtual drive, and
  adding keys to virtual UEFI BIOS to enable Secure Boot as detailed
  in ./encryption/AddingKeysToUEFIBIOS.txt

Signed-off-by: Alex Doyle <adoyle@nvidia.com>