ONIE password support example, using kvm_x86_64
If the kvm target is built with SECURE_BOOT_EXT true,
a username and password will be required for login.

For UEFI BIOS systems, the password/login will only be
required if Secure Boot is active.

This commit defaults the kvm_x86_64 virtual machine to:
Login: root
Password: onie

If Secure Boot is NOT enabled, the user will be automatically
logged in as root.

Details:
---------

New files
----------
   machine/kvm_x86_64/rootconf/sysroot-etc/passwd-open
     - passwd file configured for root login, no password
   machine/kvm_x86_64/rootconf/sysroot-etc/passwd-secured
     - passwd file configured for root login, passwd: onie
	   Note that the hash for the password has to be generated,
	   and pasted in
   machine/kvm_x86_64/rootconf/sysroot-init/open-system
     - startup script to enable/disable login if secure boot set,
	   Toggles between passwd-open and passwd-secured
   machine/kvm_x86_64/rootconf/sysroot-rcS/S12open-system
     - softlink to open-system

Build time
-----------
If SECURE_BOOT_EXT  - the secure boot extensions option
is enabled, then the ONIE password will be built in.

The password file used is copied from
machine/<manufacturer/<target>/rootconf/sysroot-etc/passwd-secured
to replace /etc/passwd in the following locations:
Rescue CD:
 /mnt/onie-boot/onie/config/etc/passwd
 /etc/passwd

Demo OS
 /etc/passwd

The machine install-platform file can override the
set_default_passwd() function in onie/installer/install.sh
to set a password on a per-machine basis.

Run time
---------

The open-system startup script will detect if the
Secure Boot UEFI variable is set or not.
If set, it sets softlinks to:
/bin/onie-console-secure
/mnt/onie-boot/onie-config/etc/passwd
(unless this is a recovery iso, in which case it
does nothing )

If Secure Boot is NOT enabled, links go to
/bin/onie-console-open
/etc/passwd-open

The resulting behavior is that, if Secure Boot is set,
the user is prompted for a login/password.
If not, pressing the enter key brings the user to a prompt.

Tests:
-------
 Build with SECURE_BOOT_EXT active in kvm_x86_64 machine.make
 Running with onie-vm.sh and UEFI BIOS (--m-uefi-bios)
  with secure boot keys installed.
 Secure boot active:
   Recovery ISO asks for password
   Embedded ONIE asks for password
 Secure boot disabled
   Recovery ISO logs in root user automatically
   Embedded ONIE logs in root user automatically

 Use onie-vm to run legacy BIOS on SECURE_BOOT_EXT image
  (In case the system BIOS is toggling between UEFI/Legacy
  modes, or a QEMU run with onie-vm does not specify
   -m-uefi-bios, and emulates legacy mode):
  Always asks for login/password

 Build without SECURE_BOOT_EXT active in kvm_x86_64 machine.make
  Build works
  Legacy BIOS - no password required for Rescue CD or Embedded ONIE
  UEFI BIOS   - no password required for Rescue CD or Embedded ONIE

Signed-off-by: Michael Shych <michaelsh@nvidia.com>
Signed-off-by: Andriy Dobush <andriyd@nvidia.com>
Signed-off-by: Alex Doyle <adoyle@nvidia.com>