UEFI Fallback recovery
If the UEFI BIOS NVRAM is corrupted, have files
present that define how an automatic recovery
should happen.

If boot entries go missing, the BIOS can detect this, and
will end up running /boot/efi/EFI/BOOT/fbx64.efi to
scan directories for BOOT.CSV files to recreate the
UEFI boot entries.

fbx64.efi is created as part of the SHIM build.
If a self-signed shim was built, and stored in
the 'safe place', a signed fbx64.efi should be there
as well, and will have been installed.

This commit updates the install configuration,
and adds BOOT.CSV files to be found for recovery.

Tests
-----
 Used onie/emulation/onie-vm.sh with Secure Boot
 active.

Install ONIE on a UEFI system
Remove recovery disk and reboot
Log in
 efibootmgr <- get ONIE boot entry number
 efibootmgr -b<ONIE boot entry number> -B  <- delete entry
 efibootmgr <- ONIE boot entry is gone
 reboot
 GRUB should come up
 log in
 efibootmgr should show ONIE boot entry number restored

Repeated with Secure Boot turned off, and it worked.

Signed-off-by: Michael Shych <michaelsh@nvidia.com>
Signed-off-by: Andriy Dobush  <andriyd@nvidia.com>
Signed-off-by: Alex Doyle <adoyle@nvidia.com>