onie-vm - support x86 and armv8 emulation runs
The QEMU machine configurations for armv8 and x86 are
very different in how they specify the cdrom and hard drives.

While it seems like they should be using the same configuration,
they are different architectures, and different BIOSs, so it
may be that there is no overlap in their Venn diagram of
functionality, and I'm not convinced it is worth the effort to
sort it out when the existing code provides two virtual machines
that behave in the same manner for the end user.

This also rolls in changes supplied by Shi Lei for the
onie-mk-iso.sh script, as well as references to a viable UEFI
BIOS to use that were part of PR #956 to address issue #940

Signed-off-by: Shi Lei <shil@centecnetworks.com>
Signed-off-by: Alex Doyle <adoyle@nvidia.com>